refactor(api-gateway): split routeHelpers from configOverrideHelpers + tighten configId
Move resolveUserIdOrSendError to routeHelpers.ts — it's a generic route
utility used by 9 files, not config-cascade-specific. Keep tryInvalidateCache
and mergeAndValidateOverrides in configOverrideHelpers.ts where they belong.

Also replace getParam with getRequiredParam for configId in both admin and
user LLM config routes. Express guarantees :id is defined on PUT/:id routes,
so the ?? '' fallback was dead defensive code that could silently mask bugs.
