+++ services/api-gateway/src/routes/user/shapes/import.ts
@@ -22,7 +22,7 @@ import { requireUserAuth } from '../../../services/AuthMiddleware.js';
 import { asyncHandler } from '../../../utils/asyncHandler.js';
 import { sendError, sendCustomSuccess } from '../../../utils/responseHelpers.js';
 import { ErrorResponses } from '../../../utils/errorResponses.js';
-import { resolveUserIdOrSendError } from '../../../utils/configOverrideHelpers.js';
+import { resolveUserIdOrSendError } from '../../../utils/routeHelpers.js';
 import { isPrismaUniqueConstraintError } from '../../../utils/prismaErrors.js';
 import type { AuthenticatedRequest } from '../../../types.js';
